refactor(cloud-workers): split quiescence renewal script
